% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get tickers data}
\usage{
getData()
}
\arguments{
\item{tickers}{}
}
\value{
Tickers Data
}
\description{
Get tickers data based on suggested list of tickers
}
\examples{
getData()

}
